Reconfigure buzzer
Turns out a prescaler of 2 is enough to cover most notes including some
notes in octave 0 and above. This is enough to cover all notes in the
standard piano.

A macro file is renamed `period_macros.h` to `note_macros.h`. All the
period values, clock selection for relative files have been changed.