Implement Flash File Read System
Created files are:
    `include/data.h`
    `include/flash.h`
    `src/flash.c`

There's actually a `src/data.c` file, which will be automatically
generated through a (Python) script that's prerun before compilation.
This will be clearer in future commits. Hence, `src/data.c` is added in
.gitignore.

The main implementation of the file system is in `src/flash.c` and
`include/flash.h`. `src/data.c` just contains the contents of the files.