Implement block validation
Created `scripts/validate_blocks.py` and `scripts/constants.py`.
The term 'block' refers to the pairs of brackets '[]' and '{}'. '[]' are
bars (as in music bars) and '{}' are called setting blocks. These terms
are widely used in the `scripts/validate_blocks.py` file and will
continue to do so.

The `scripts/constants.py` file currently has constants for
`scripts/validate_blocks.py` but will have more constants for other
potential Python files in the future.

The .gitignore file is modified to ignore the `scripts/__pycache__`
directory for implementation testing purposes.