Organize Files in scripts/
The validation-related files in scripts/ are mostly renamed. Functions
are now categorized more explicitly according to their file names and
are pretty much self-explanatory.

The files created are:
    `validate_bar_content.py`
    `validate_block_enclosures.py`
    `validate_setting_content.py`

As of now, the `validate_blocks.py` file is the main file to validate
all blocks. I'm guessing that this last long.

In `scripts/constants`, some files are renamed as well to now become:
    `block_enclosures.py`
    `notes.py`
    `setting_fields.py`

All validation files have imported the `constants/` package correctly.