Partially Implement Overall Sheet Validation
Have implemented all of the three types of validation in
`scripts/process_sheets.py` but not complete and/or satisfied yet.
The code is very very ugly, I will plan to refactor them into their own
separate functions rather than having all of the implementation there.
Since, this is not done, it still generates `src/data.c` as it is.

In `scripts/validate_block_enclosures.py` it now returns a Boolean
rather than the error count. `sheets/example_hello.txt` is slightly
modified for testing purposes.