Implement Slur Checking and Improve Error Messages
Finally implemented slur checking, where slurs must always open and
close right after and cannot be reopend once it's already open or close
once it's already closed. I decided that at the end of the sheet to not
check if it's still opened.

Main changes are in `get_note_beats` and `get_tuplet_beats` of
`scripts/bar_helper.py`. `scripts/validate_blocks.py` just keeps track
of the final slur state and passes it to the function as well.

Error messages when ElementError is raised are now more specific. This
requires adding a `msg` attribute in `ElementError`, in the end it's
very similar to FieldError.

`sheets/example_hello.txt` is edited slightly yet again for testing
purposes.