Bar Helper More Explicit
`scripts/bar_helper.py` needs to be more explicit with what info it is
returning for implementing music code in the future. In this commit
all the functions are made to return including:
    - its note/symbol
    - duration in 32nd
    - note additionals if any

For tuplets, it returns a list of tuples. The files that is mainly
affected is `scripts/sheet_processor.py`.

The `scripts/process_sheets.py` file is slightly formatted for
successful validations :)