Implement Music Code Generation for Sheets
Created `scripts/generate_music_code.py` that has `generate_music_code`
which creates music code for the entire sheet.

Functions in `scripts/music_code_helper.py` are slightly modified to now
also return the total ticks. This is useful for handling anacrusises.

`scripts/process_sheets.py` is modified to write generated music code
to a directory `/music code`. As of now, the file name of the output
are the same as the sheet's filename.

`.gitignore` is modified to ignore `music code/`.