Reformat Music Codes Per Line
Previously the `music_codes_per_line` variable in
`scripts/process_sheets.py` correctly formats the resulted music codes
to a single string separated with spaces and newlines but at the very
end of the string, a newline is always appended. This is fixed so that
the newline at the end is never appended.

Additionally, all example error sheets are removed:
    `/sheets/example_break_errors.txt`
    `/sheets/example_hello.txt`
    `/sheets/example_note_errors.txt`
    `/sheets/example_tuplet_errors.txt`