Fix DP and Led Timing Issue when Music Playing
Now timing when the DP is supposed to alternate and increment bar count
seems fixed, however still needs checking. Another slight issue is when
the bar count is supposed to increment, it's supposed to increment on
the first beat not the last one. I hope I get this fixed on future
commits.

The main way to solve the issue was to immediately overwrite
`ticks_play` and `ticks_break` right when both of them are 0 and not
waiting for the next interrupt for it to be finally overwritten.