Fix Bar Increment Timing When Music Playing
Turns out, when `bar_count` is incremented isn't at the right timing.
It increments at the last beat, for example at a 4/4 time signature,
it increments at the fourth beat and not the first beat. This issue
is solved by initially setting `bar_counter` to `tsig_top` - 1. A
similar approach is used in anacrusises as well which also solves the
issue.