State Machine for Music Interactivity
State machine in `src/main.c` has been implemented now allowing for
music to be selected, paused, played, stopped. Here's a brief
explanation on how the state machine works:

Initially, the number displayed is the sheet number, the user can press
S1 or S2 to decrement or increment the number respectively. To make a
selection, S3 is pressed. To start playing the music, S4 can be pressed.
However, any other button would result in a cancellation of the
selection made.

At this stage, the music is now being played. While it's playing, it can
be paused by pressing S3. After paused, the music can be stopped or
resumed playing by either pressing S2 or S4 respectively. If resumed,
then it'll continue as normal, but if stopped, it'll go back to the
selection stage.

Additional changes in `src/music.c` is added to allow the music to be
paused/played properly. Specifically in `music_pause` and `music_play`,
where the last played note has to be stored before silenced.