Better Pointers for Error Messages
Rather than having the error message always after the pointer, now
depending if the length of the error message is short enough, the error
message is now before the pointer.

This is done by having a helper function `construct_error_msg`. All
parts of the code that are relevant for creating these error messages
now utilizes this function.