Music Code Generation Bug With Skip Bars Fix
There's a bug where an error might be raised when dealing with skip bars
and slurs. During music code generation, the slur states are never
updated. So there is potential for it to skip to a bar that updates the
slur state.

This bug is fixed by handling the `skip_bars` variable after the bar
has been processed. Additionally, anacrusis is always overwritten even
when there are bars to skip.