Rename Break to Rest
All this time I just realized that the correct term for a period of
silence in music is 'rest'. All this time I thought it was `break` which
isn't as correct.

All files that uses the term 'break' whether it be variables, or
comments have been renamed to use 'rest'. The affected files are:
    `include/macros/music_macros.h`
    `scripts/bar_helper.py`
    `scripts/constants/music_code.py`
    `scripts/constants/notes.py`
    `scripts/generate_music_code.py`
    `scripts/music_code_helper.py`
    `scripts/process_sheets.py`
    `scripts/validate_sheet.py`
    `src/music.c`