Add README Files
I finally created README files. There is a README file in the root
directory, the `sheets/` directory and in the `music code/` directory.
The `music code/` directory is no longer ignored.

Additionally, the `scripts/process_sheets.py` file must be modified a
little bit to prevent it from deleting everything in the `music code/`
directory.