Change BPM Syntax
I just realized that BPMs don't always have a quarter note be a beat.
The beat value may be an eighth note, a dotted quarter note, etc. Now,
to write a BPM, the value is two integers separated with a '/'. The
first integer is the desired BPM value and the second integer should be
1, 2, 4, 8, 16 or 32. It can optionally be followed by a '.' to indicate
that the beat value is a dotted note.

The BPM's music code is also changed, right after the BPM's prefix would
immediately be the period value for the TCB. There's no more period
calculation during runtime, just parsing.

The main code changes are in:
    `scripts/constants/music_code.py`
    `scripts/constants/setting_fields.py`
    `scripts/generate_music_code.py`
    `scripts/setting_block_helper.py`
    `src/music.c`

Mainly the changes are in `scripts/setting_block_helper.py` in the
`field_to_key_val` function. `src/music.c` has a simple change which is
to immediately parse the value into `next_bpm_per`.