Fix BPM Miscalculation for Dotted Note Beats
There was a miscalculation for calcalating BPM's period value in a case
where the beat value is dotted. The right calculation is to convert a
bpm value to a bpm value where the beat size is a quarter note is:

    bpm_in_quarter_note = bpm_value * 4 / beat_value
    if dotted:
        bpm_in_quarter_note *= 1.5

The right calculation has been implemented, main changes are in
`scripts/setting_block_helper.py`.